OverlandFlow class rename & easy import
…in line with other components and discussion on this commit
